# optimize the `medium` algorithm