--- a/fuzz/fuzz_targets/checksum.rs
+++ b/fuzz/fuzz_targets/checksum.rs
@@ -21,23 +21,6 @@ fuzz_target!(|input: (Vec<u8>, u32)| {
     }
 
     {
-        let expected = {
-            let mut h = crc32fast::Hasher::new_with_initial(0);
-            h.update(&input[..]);
-            h.finalize()
-        };
-
-        let mut buf = [0; 1 << 16];
-        let mut dst = zlib_rs::read_buf::ReadBuf::new(&mut buf[..input.len()]);
-
-        let actual = zlib_rs::crc32::crc32_copy(&mut dst, input.as_slice());
-
-        assert_eq!(expected, actual);
-
-        assert_eq!(input, dst.filled());
-    }
-
-    {
         use zlib_rs::{crc32, crc32_combine};
 
         let data = &input;