--- a/zlib-rs/src/deflate/algorithm/medium.rs
+++ b/zlib-rs/src/deflate/algorithm/medium.rs
@@ -211,6 +211,7 @@ fn emit_match(state: &mut State, mut m: Match) -> bool {
     bflush
 }
 
+#[inline(always)]
 fn insert_match(state: &mut State, mut m: Match) {
     if state.lookahead <= (m.match_length as usize + WANT_MIN_MATCH) {
         return;