@bjorn3 2024-10-14
Opened https://github.com/trifectatechfoundation/zlib-rs/pull/224 against this PR to fix CI.